baselines/BAF/notebooks/requirements.txt
aequitas==0.42.0
ctgan==0.4.3
lightgbm==3.1.1
matplotlib==3.3.4
# numpy==1.20.2
numpy
optuna==2.10.0
pandas==1.1.4
PyYAML==5.4.1
scikit-learn==0.24.2
# scipy==1.6.2
scipy
seaborn==0.11.1